change i8 to i16

+++ protocol/src/packet/serverbound/play.rs
@@ -71,5 +71,5 @@ pub struct PlayerBlockPlacementPacket {
 
 #[derive(PacketDef)]
 pub struct HeldItemChangePacket {
-    pub slot: i8
+    pub slot: i16
 }
